add rarity weighting

+++ src/wavespawn.rs
@@ -4,8 +4,8 @@ pub struct Wavespawn {
     //Information loaded from configs
     pub squads: Vec<Bot>,
     pub tags: Vec<String>,
-    pub weight: i64,
-    pub rarity: i64,
+    pub weight: f64,
+    pub rarity: f64,
     pub spawn_tank: bool,
 }
